+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- [x] Updated Packages - apparently no side-effects this time \o/ Implemeneted Tooling for Vue Components, reconfigured chunks so app is always needed, admin and empire|fleets|etc are added via pug template. `05.11.2017`
 - [x] externalize cron/turn handling into a seperate node process. `01.10.2017`
 - [x] add orbital index to planets `01.10.2017`
 - [x] enlist => assign random player system `01.10.2017`